+++ CHANGELOG.md
@@ -52,3 +52,4 @@
 - Bug #48: Update exception messages for `BaseBlockElement` and `BaseInlineElement` class to include empty tag name (@terabytesoftw)
 - Bug #49: Update exception message handling in `InlineElementTest` and clarify test case descriptions in `BlockElementProvider` class (@terabytesoftw)
 - Bug #50: Better naming for `BlockContent`, `InlineContent`, and `VoidContent` classes and update related tests (@terabytesoftw)
+- Bug #51: Normalize tag names to lowercase in `BaseBlockElement`, `BaseInlineElement`, and `BaseVoidElement` classes; update related tests and data providers (@terabytesoftw)